+++ test/hello-world/default.nix
@@ -1,10 +1,10 @@
 {
   # Package sources:
   sources ? import ../../nix/sources.nix
-, # nixpkgs:
-  pkgs ? import sources.nixpkgs { }
-, # Which version of GHC to use:
-  compiler ? "default"
+  # nixpkgs:
+, pkgs ? import sources.nixpkgs { }
+  # Which version of GHC to use:
+, compiler ? "default"
 }:
 let nix-hs = import ../../default.nix { inherit pkgs; };